+++ src/app-server/routes/index.test.js
@@ -1,5 +1,5 @@
-import getRoutes from './routes';
-import { getServer } from './util/testUtils';
+import getRoutes from './';
+import { getServer } from '../../util/testUtils';
 
 function getResponse(injectRequest, server = getServer()) {
 	// a Promise that returns the server instance after it has been
